fix: disable WithSenseClientDisconnection for windows
